refactoring: #845 add and use convenience import for module "status"
