+++ osmaxx/conversion/__init__.py
@@ -1,5 +1,5 @@
-from .constants import output_format
+from .constants import output_format, status
 
 default_app_config = 'osmaxx.conversion.apps.ConversionConfig'
 
-__all__ = ['default_app_config', 'output_format']
+__all__ = ['default_app_config', 'output_format', 'status']
