+++ osmaxx/conversion/models.py
@@ -12,7 +12,7 @@ from osmaxx.clipping_area.models import ClippingArea
 from osmaxx.conversion.converters.converter import convert
 from osmaxx.conversion.converters.converter_gis.detail_levels import DETAIL_LEVEL_CHOICES, DETAIL_LEVEL_ALL
 from osmaxx.conversion.constants.coordinate_reference_systems import CRS_CHOICES
-from osmaxx.conversion.constants import status
+from osmaxx.conversion import status
 
 
 def job_directory_path(instance, filename):
