+++ osmaxx/excerptexport/models/export.py
@@ -9,7 +9,7 @@ from django.utils.translation import ugettext_lazy as _
 from rest_framework.reverse import reverse
 
 from osmaxx.conversion import output_format
-from osmaxx.conversion.constants import status
+from osmaxx.conversion import status
 from osmaxx.excerptexport._settings import RESULT_FILE_AVAILABILITY_DURATION, EXTRACTION_PROCESSING_TIMEOUT_TIMEDELTA
 
 logger = logging.getLogger(__name__)
